chore: revert breaking change
