+++ packages/discord.js/src/managers/GuildMessageManager.js
@@ -1,8 +1,6 @@
 'use strict';
 
-const { Routes } = require('discord-api-types/v10');
 const MessageManager = require('./MessageManager');
-const { DiscordjsTypeError, ErrorCodes } = require('../errors');
 
 /**
  * Manages API methods for messages in a guild and holds their cache.
@@ -14,19 +12,6 @@ class GuildMessageManager extends MessageManager {
    * @name GuildMessageManager#channel
    * @type {GuildTextBasedChannel}
    */
-
-  /**
-   * Publishes a message in an announcement channel to all channels following it, even if it's not cached.
-   * @param {MessageResolvable} message The message to publish
-   * @returns {Promise<Message>}
-   */
-  async crosspost(message) {
-    message = this.resolveId(message);
-    if (!message) throw new DiscordjsTypeError(ErrorCodes.InvalidType, 'message', 'MessageResolvable');
-
-    const data = await this.client.rest.post(Routes.channelMessageCrosspost(this.channel.id, message));
-    return this.cache.get(data.id) ?? this._add(data);
-  }
 }
 
 module.exports = GuildMessageManager;
